overtime logic on detail data timesheet get service, app.yaml set max-instances and min instances to lower from before, fix error when deploying

+++ src/main/java/com/treemaswebapi/treemaswebapi/service/DetailData/Timesheet/TimesheetResponse.java
@@ -26,6 +26,6 @@ public class TimesheetResponse {
     LocalTime jamMsk;
     LocalTime jamPlg;
     BigDecimal totalJamKerja;
-    BigDecimal overtime;
+    int overtime;
     String notePekerjaan;
 }
